pyproject.toml
[project]
name = "labprojetos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.12.*"
dependencies = [
    "flask>=3.1.0",
    "keras>=3.9.1",
    "matplotlib>=3.10.1",
    "pandas>=2.2.3",
    "pygame>=2.6.1",
    "pyqt6>=6.8.1",
    "pyserial>=3.5",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.2",
    "tensorflow>=2.19.0",
    "typer>=0.15.2",
]
